+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # rdbus
 
+- `rdbus` now outputs an error if incoming NMEA data could not be parsed in any way (trash).
 - Added check that validates that known `"data_type"` is not present with unnecessary `"data_order"`.
 
 # 0.2.0